# docs/swagger.yaml
basePath: /api
definitions:
  models.Message:
    properties:
      message:
        type: string
    type: object
  models.Task:
    properties:
      created:
        type: string
      description:
        type: string
      id:
        type: integer
      is_finished:
        type: boolean
      title:
        type: string
    type: object
info:
  contact:
    email: alexander.klonov@mail.ru
  description: Test Task for building api for todolist
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ToDoList api
  version: "1.0"
paths:
  /task/{title}:
    delete:
      parameters:
      - description: Task title
        in: path
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Message'
      summary: Delete existing task
    get:
      parameters:
      - description: Task title
        in: path
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      summary: Get one task by its name
    post:
      parameters:
      - description: Task title
        in: path
        name: title
        required: true
        type: string
      - description: Param to change task status
        in: query
        name: finished
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      summary: Change task status
  /task/create:
    post:
      parameters:
      - description: Task title
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: Task full description
        in: body
        name: description
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Message'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.Message'
      summary: Creates new task
  /tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      summary: Get all tasks
swagger: "2.0"